+++ services/admin/src/routes/object/edit/index.svelte
@@ -27,8 +27,6 @@
   }
 </script>
 
-<br />
-<br />
 <PrefetchLoader bind:href={objectHref}>
   <div class="search">
     <TypeAhead
